# docs/swagger.yaml
basePath: '{{.BasePath}}'
definitions:
  models.Response:
    properties:
      code:
        example: 200
        type: integer
      data:
        type: object
      msg:
        example: ok
        type: string
    type: object
host: '{{.Host}}'
info:
  contact: {}
  description: '{{.Description}}'
  license: {}
  title: '{{.Title}}'
  version: '{{.Version}}'
paths:
  /patients:
    get:
      consumes:
      - application/json
      parameters:
      - description: 医生id
        in: query
        name: doctor_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
            type: object
      summary: 病人列表
  /user:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户名
        in: query
        name: username
        required: true
        type: string
      - description: 密码
        in: query
        name: password
        required: true
        type: string
      - description: 用户类型
        in: query
        name: user_type
        required: true
        type: string
      - description: 姓名
        in: query
        name: name
        type: string
      - description: 头像
        in: query
        name: avatar
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
            type: object
      summary: 注册用户
  /user/login:
    post:
      consumes:
      - application/json
      description: get a token string
      parameters:
      - description: 用户名
        in: query
        name: username
        required: true
        type: string
      - description: 密码
        in: query
        name: password
        required: true
        type: string
      - description: 用户类型
        in: query
        name: user_type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
            type: object
      summary: 普通用户登录
swagger: "2.0"